Address Grok-1 coverage review findings (raulmc-5kx, #24)
