+++ src/inventory/mod.rs
@@ -12,28 +12,21 @@
 // Deeper semantic analysis (routing math, expert-level statistics,
 // dequantized parameter accounting) is explicitly *not* done here.
 
-use std::collections::{BTreeMap, BTreeSet};
 use std::path::{Path, PathBuf};
 
 use anyhow::{Context, Result, bail};
 
 use crate::parser::{self, RawTensor};
 use crate::schema::{
-    BlockSummary, Grok1CoverageCounts, Grok1CoverageManifest, Grok1UnknownSlot,
-    InferredHyperparams, InventoryTotals, KindCount, ModelInventory, MoeProjection,
+    BlockSummary, InferredHyperparams, InventoryTotals, KindCount, ModelInventory, MoeProjection,
     QuantizedAttentionWidth, ShardRange, TensorDType, TensorInfo, TensorKind, TensorRole,
 };
 
-pub const SCHEMA_VERSION: u32 = 2;
-pub const GROK1_COVERAGE_SCHEMA_VERSION: u32 = 1;
+mod grok1_coverage;
 
-const GROK1_EXPECTED_BLOCKS: u32 = 64;
-const GROK1_EXPECTED_TENSORS: u64 = 770;
-const GROK1_EXPECTED_ROUTERS: u64 = 64;
-const GROK1_EXPECTED_EXPERT_FAMILIES: u64 = 64 * 3;
-const GROK1_D_MODEL: u64 = 6_144;
-const GROK1_D_FF: u64 = 32_768;
-const GROK1_N_EXPERTS: u64 = 8;
+pub use grok1_coverage::{should_validate_grok1_coverage, validate_grok1_complete_manifest};
+
+pub const SCHEMA_VERSION: u32 = 2;
 
 /// Configuration for enumerating and classifying a checkpoint.
 #[derive(Clone, Debug)]
@@ -133,101 +126,6 @@ pub fn build_inventory(path: &Path, cfg: &InventoryConfig) -> Result<ModelInvent
         totals,
         schema_version: SCHEMA_VERSION,
     })
-}
-
-/// Validate a complete Grok-1 inventory and emit a deterministic coverage
-/// manifest. This is intentionally stricter than generic inventory building:
-/// complete Grok-1 manifests must fail closed on missing blocks/tensors,
-/// duplicate structural names, unknown/unexpected layouts, or incomplete
-/// hyperparameter metadata.
-pub fn validate_grok1_complete_manifest(inv: &ModelInventory) -> Result<Grok1CoverageManifest> {
-    if inv.model_family != "grok-1" {
-        bail!("Grok-1 coverage validation requires model_family 'grok-1'");
-    }
-
-    let expected = Grok1CoverageCounts {
-        blocks: GROK1_EXPECTED_BLOCKS,
-        tensors: GROK1_EXPECTED_TENSORS,
-        routers: GROK1_EXPECTED_ROUTERS,
-        expert_families: GROK1_EXPECTED_EXPERT_FAMILIES,
-        unknown_tensors: 0,
-    };
-    let discovered = grok1_discovered_counts(inv);
-    let unknown_slots = grok1_unknown_slots(inv);
-    let checksum = grok1_checksum(inv, &discovered, &unknown_slots);
-
-    let manifest = Grok1CoverageManifest {
-        model_family: inv.model_family.clone(),
-        schema_version: inv.schema_version,
-        coverage_schema_version: GROK1_COVERAGE_SCHEMA_VERSION,
-        validation: "pass".to_string(),
-        checksum,
-        expected: expected.clone(),
-        discovered: discovered.clone(),
-        unknown_slots,
-    };
-
-    let mut errors = Vec::new();
-    if inv.schema_version != SCHEMA_VERSION {
-        errors.push(format!(
-            "schema_version {} != expected {}",
-            inv.schema_version, SCHEMA_VERSION
-        ));
-    }
-    if inv.inferred.n_blocks != Some(GROK1_EXPECTED_BLOCKS) {
-        errors.push(format!(
-            "inferred n_blocks {:?} != expected {}",
-            inv.inferred.n_blocks, GROK1_EXPECTED_BLOCKS
-        ));
-    }
-    if inv.inferred.d_model != Some(GROK1_D_MODEL) {
-        errors.push(format!(
-            "inferred d_model {:?} != expected {}",
-            inv.inferred.d_model, GROK1_D_MODEL
-        ));
-    }
-    if inv.inferred.d_ff != Some(GROK1_D_FF) {
-        errors.push(format!(
-            "inferred d_ff {:?} != expected {}",
-            inv.inferred.d_ff, GROK1_D_FF
-        ));
-    }
-    if inv.inferred.n_experts != Some(GROK1_N_EXPERTS) {
-        errors.push(format!(
-            "inferred n_experts {:?} != expected {}",
-            inv.inferred.n_experts, GROK1_N_EXPERTS
-        ));
-    }
-
-    compare_count("blocks", discovered.blocks, expected.blocks, &mut errors);
-    compare_count("tensors", discovered.tensors, expected.tensors, &mut errors);
-    compare_count("routers", discovered.routers, expected.routers, &mut errors);
-    compare_count(
-        "expert_families",
-        discovered.expert_families,
-        expected.expert_families,
-        &mut errors,
-    );
-    compare_count(
-        "unknown_tensors",
-        discovered.unknown_tensors,
-        expected.unknown_tensors,
-        &mut errors,
-    );
-
-    validate_grok1_blocks(inv, &mut errors);
-    validate_unique_source_keys(inv, &mut errors);
-    validate_unique_structural_names(inv, &mut errors);
-    validate_grok1_expected_slots(inv, &mut errors);
-
-    if !errors.is_empty() {
-        bail!(
-            "Grok-1 complete manifest validation failed: {}",
-            errors.join("; ")
-        );
-    }
-
-    Ok(manifest)
 }
 
 // --- Shard enumeration -----------------------------------------------------
@@ -585,7 +483,7 @@ fn grok_layout_candidate(
 
 // --- Block summaries -------------------------------------------------------
 
-fn summarize_blocks(tensors: &[TensorInfo]) -> Vec<BlockSummary> {
+pub(crate) fn summarize_blocks(tensors: &[TensorInfo]) -> Vec<BlockSummary> {
     use std::collections::BTreeMap;
 
     // Bucket tensors: None => embedding + final norm singletons, Some(i) => block i.
@@ -724,320 +622,6 @@ fn compute_totals(tensors: &[TensorInfo]) -> InventoryTotals {
         }
     }
     out
-}
-
-fn grok1_discovered_counts(inv: &ModelInventory) -> Grok1CoverageCounts {
-    let block_indices = inv
-        .tensors
-        .iter()
-        .filter_map(|tensor| tensor.block_index)
-        .collect::<BTreeSet<_>>();
-
-    Grok1CoverageCounts {
-        blocks: block_indices.len() as u32,
-        tensors: inv.tensors.len() as u64,
-        routers: inv
-            .tensors
-            .iter()
-            .filter(|tensor| matches!(tensor.kind, TensorKind::Router))
-            .count() as u64,
-        expert_families: inv
-            .tensors
-            .iter()
-            .filter(|tensor| matches!(tensor.kind, TensorKind::MoeExpertProjection { .. }))
-            .count() as u64,
-        unknown_tensors: inv
-            .tensors
-            .iter()
-            .filter(|tensor| matches!(tensor.kind, TensorKind::Unknown { .. }))
-            .count() as u64,
-    }
-}
-
-fn grok1_unknown_slots(inv: &ModelInventory) -> Vec<Grok1UnknownSlot> {
-    let mut out = inv
-        .tensors
-        .iter()
-        .filter_map(|tensor| match &tensor.kind {
-            TensorKind::Unknown { reason } => Some(Grok1UnknownSlot {
-                structural_name: grok1_structural_name(tensor),
-                block_index: tensor.block_index,
-                block_slot: tensor.block_slot,
-                shape: tensor.shape.clone(),
-                reason: reason.clone(),
-            }),
-            _ => None,
-        })
-        .collect::<Vec<_>>();
-    out.sort_by(|a, b| a.structural_name.cmp(&b.structural_name));
-    out
-}
-
-fn compare_count<T>(label: &str, actual: T, expected: T, errors: &mut Vec<String>)
-where
-    T: Copy + Eq + std::fmt::Display,
-{
-    if actual != expected {
-        errors.push(format!("{label} {actual} != expected {expected}"));
-    }
-}
-
-fn validate_grok1_blocks(inv: &ModelInventory, errors: &mut Vec<String>) {
-    let mut by_block: BTreeMap<u32, Vec<&TensorInfo>> = BTreeMap::new();
-    for tensor in &inv.tensors {
-        if let Some(block_index) = tensor.block_index {
-            by_block.entry(block_index).or_default().push(tensor);
-        }
-    }
-
-    for block_index in 0..GROK1_EXPECTED_BLOCKS {
-        match by_block.get(&block_index) {
-            Some(tensors) if tensors.len() == 12 => {}
-            Some(tensors) => errors.push(format!(
-                "block_{block_index:03} has {} tensors, expected 12",
-                tensors.len()
-            )),
-            None => errors.push(format!("missing block_{block_index:03}")),
-        }
-    }
-
-    for block_index in by_block.keys() {
-        if *block_index >= GROK1_EXPECTED_BLOCKS {
-            errors.push(format!("unexpected block_{block_index:03}"));
-        }
-    }
-
-    for tensor in &inv.tensors {
-        if tensor.block_index.is_none()
-            && !matches!(
-                tensor.kind,
-                TensorKind::TokenEmbedding | TensorKind::FinalNorm
-            )
-        {
-            errors.push(format!(
-                "unexpected unassigned tensor {}",
-                grok1_structural_name(tensor)
-            ));
-        }
-    }
-}
-
-fn validate_unique_structural_names(inv: &ModelInventory, errors: &mut Vec<String>) {
-    let mut seen = BTreeSet::new();
-    for tensor in &inv.tensors {
-        let name = grok1_structural_name(tensor);
-        if !seen.insert(name.clone()) {
-            errors.push(format!("duplicate structural name {name}"));
-        }
-    }
-}
-
-fn validate_unique_source_keys(inv: &ModelInventory, errors: &mut Vec<String>) {
-    let mut seen = BTreeSet::new();
-    for tensor in &inv.tensors {
-        let key = (tensor.shard_ordinal, tensor.in_shard_index);
-        if !seen.insert(key) {
-            errors.push(format!(
-                "duplicate tensor key shard={} in_shard={}",
-                tensor.shard_ordinal, tensor.in_shard_index
-            ));
-        }
-    }
-}
-
-fn validate_grok1_expected_slots(inv: &ModelInventory, errors: &mut Vec<String>) {
-    for block_index in 0..GROK1_EXPECTED_BLOCKS {
-        for slot in 0..12u32 {
-            let tensors = inv
-                .tensors
-                .iter()
-                .filter(|tensor| {
-                    tensor.block_index == Some(block_index) && tensor.block_slot == Some(slot)
-                })
-                .collect::<Vec<_>>();
-            if tensors.is_empty() {
-                errors.push(format!("missing block_{block_index:03}.slot_{slot:02}"));
-            }
-            if tensors.len() > 1 {
-                errors.push(format!(
-                    "duplicate block_{block_index:03}.slot_{slot:02} entries: {}",
-                    tensors.len()
-                ));
-            }
-            if let Some(tensor) = tensors.first() {
-                validate_grok1_slot_tensor(tensor, errors);
-            }
-        }
-    }
-}
-
-fn validate_grok1_slot_tensor(tensor: &TensorInfo, errors: &mut Vec<String>) {
-    let Some(slot) = tensor.block_slot else {
-        return;
-    };
-    let name = grok1_structural_name(tensor);
-    match slot {
-        0 | 2 => validate_tensor_signature(
-            tensor,
-            &name,
-            TensorRole::QuantWeight,
-            TensorDType::I8,
-            &[GROK1_N_EXPERTS, GROK1_D_MODEL, GROK1_D_FF],
-            |kind| matches!(kind, TensorKind::MoeExpertProjection { .. }),
-            errors,
-        ),
-        1 => validate_tensor_signature(
-            tensor,
-            &name,
-            TensorRole::QuantWeight,
-            TensorDType::I8,
-            &[GROK1_N_EXPERTS, GROK1_D_FF, GROK1_D_MODEL],
-            |kind| matches!(kind, TensorKind::MoeExpertProjection { .. }),
-            errors,
-        ),
-        3 | 6 => validate_tensor_signature(
-            tensor,
-            &name,
-            TensorRole::QuantWeight,
-            TensorDType::I8,
-            &[GROK1_D_MODEL, 1_024],
-            |kind| {
-                matches!(
-                    kind,
-                    TensorKind::QuantizedAttentionProjection {
-                        width: QuantizedAttentionWidth::Narrow
-                    }
-                )
-            },
-            errors,
-        ),
-        4 | 5 => validate_tensor_signature(
-            tensor,
-            &name,
-            TensorRole::QuantWeight,
-            TensorDType::I8,
-            &[GROK1_D_MODEL, GROK1_D_MODEL],
-            |kind| {
-                matches!(
-                    kind,
-                    TensorKind::QuantizedAttentionProjection {
-                        width: QuantizedAttentionWidth::ModelWidth
-                    }
-                )
-            },
-            errors,
-        ),
-        11 => validate_tensor_signature(
-            tensor,
-            &name,
-            TensorRole::Tensor,
-            TensorDType::F32,
-            &[GROK1_D_MODEL, GROK1_N_EXPERTS],
-            |kind| matches!(kind, TensorKind::Router),
-            errors,
-        ),
-        7..=10 => {
-            if matches!(tensor.kind, TensorKind::Unknown { .. }) {
-                errors.push(format!("unexpected unknown tensor at {name}"));
-            }
-        }
-        _ => errors.push(format!("unexpected block slot {slot} at {name}")),
-    }
-}
-
-fn validate_tensor_signature(
-    tensor: &TensorInfo,
-    name: &str,
-    role: TensorRole,
-    dtype: TensorDType,
-    shape: &[u64],
-    kind_matches: impl FnOnce(&TensorKind) -> bool,
-    errors: &mut Vec<String>,
-) {
-    if tensor.role != role
-        || tensor.dtype != dtype
-        || tensor.shape.dims() != shape
-        || !kind_matches(&tensor.kind)
-    {
-        errors.push(format!(
-            "unexpected tensor layout at {name}: role={} dtype={} shape={} kind={}",
-            tensor.role.label(),
-            tensor.dtype.label(),
-            tensor.shape.render(),
-            tensor.kind.short_label()
-        ));
-    }
-}
-
-fn grok1_structural_name(tensor: &TensorInfo) -> String {
-    match (tensor.block_index, tensor.block_slot) {
-        (Some(block), Some(slot)) => {
-            format!(
-                "block_{block:03}.slot_{slot:02}.{}",
-                tensor.kind.short_label()
-            )
-        }
-        (Some(block), None) => format!(
-            "block_{block:03}.slot_unknown.{}",
-            tensor.kind.short_label()
-        ),
-        (None, _) if matches!(tensor.kind, TensorKind::TokenEmbedding) => "embedding".to_string(),
-        (None, _) if matches!(tensor.kind, TensorKind::FinalNorm) => "final_norm".to_string(),
-        (None, Some(slot)) => format!("unassigned.slot_{slot:02}.{}", tensor.kind.short_label()),
-        (None, None) => format!("unassigned.{}", tensor.kind.short_label()),
-    }
-}
-
-fn grok1_checksum(
-    inv: &ModelInventory,
-    discovered: &Grok1CoverageCounts,
-    unknown_slots: &[Grok1UnknownSlot],
-) -> String {
-    let mut lines = vec![
-        format!("model_family={}", inv.model_family),
-        format!("schema_version={}", inv.schema_version),
-        format!("blocks={}", discovered.blocks),
-        format!("tensors={}", discovered.tensors),
-        format!("routers={}", discovered.routers),
-        format!("expert_families={}", discovered.expert_families),
-        format!("unknown_tensors={}", discovered.unknown_tensors),
-    ];
-
-    let mut tensor_lines = inv
-        .tensors
-        .iter()
-        .map(|tensor| {
-            format!(
-                "tensor|{}|{}|{}|{}|{}|{}|{}",
-                grok1_structural_name(tensor),
-                tensor.shard_ordinal,
-                tensor.in_shard_index,
-                tensor.role.label(),
-                tensor.dtype.label(),
-                tensor.shape.render(),
-                tensor.kind.short_label()
-            )
-        })
-        .collect::<Vec<_>>();
-    tensor_lines.sort();
-    lines.extend(tensor_lines);
-    for unknown in unknown_slots {
-        lines.push(format!(
-            "unknown|{}|{}",
-            unknown.structural_name, unknown.reason
-        ));
-    }
-
-    stable_fnv1a64(lines.join("\n").as_bytes())
-}
-
-fn stable_fnv1a64(bytes: &[u8]) -> String {
-    let mut hash = 0xcbf2_9ce4_8422_2325u64;
-    for byte in bytes {
-        hash ^= u64::from(*byte);
-        hash = hash.wrapping_mul(0x0000_0100_0000_01b3);
-    }
-    format!("fnv1a64:{hash:016x}")
 }
 
 #[cfg(test)]
@@ -1208,98 +792,6 @@ mod tests {
         assert_eq!(narrow.short_label(), "attn_proj_i8.narrow");
     }
 
-    #[test]
-    fn validates_complete_grok1_coverage_manifest() {
-        let inv = complete_grok1_inventory();
-
-        let manifest = validate_grok1_complete_manifest(&inv).expect("valid manifest");
-
-        assert_eq!(manifest.expected.tensors, 770);
-        assert_eq!(manifest.discovered.tensors, 770);
-        assert_eq!(manifest.expected.blocks, 64);
-        assert_eq!(manifest.discovered.blocks, 64);
-        assert_eq!(manifest.expected.routers, 64);
-        assert_eq!(manifest.discovered.routers, 64);
-        assert_eq!(manifest.expected.expert_families, 192);
-        assert_eq!(manifest.discovered.expert_families, 192);
-        assert!(manifest.unknown_slots.is_empty());
-        assert_eq!(manifest.checksum, "fnv1a64:ce4d8e7cde002f74");
-    }
-
-    #[test]
-    fn grok1_coverage_fails_on_missing_block() {
-        let mut inv = complete_grok1_inventory();
-        inv.tensors.retain(|tensor| tensor.block_index != Some(10));
-        refresh_inventory_derived_fields(&mut inv);
-
-        let err = validate_grok1_complete_manifest(&inv).unwrap_err();
-
-        assert!(format!("{err:#}").contains("missing block_010"));
-    }
-
-    #[test]
-    fn grok1_coverage_fails_on_missing_tensor() {
-        let mut inv = complete_grok1_inventory();
-        inv.tensors
-            .retain(|tensor| !(tensor.block_index == Some(7) && tensor.block_slot == Some(4)));
-        refresh_inventory_derived_fields(&mut inv);
-
-        let err = validate_grok1_complete_manifest(&inv).unwrap_err();
-
-        assert!(format!("{err:#}").contains("missing block_007.slot_04"));
-    }
-
-    #[test]
-    fn grok1_coverage_fails_on_duplicate_tensor_key() {
-        let mut inv = complete_grok1_inventory();
-        let duplicate = inv
-            .tensors
-            .iter()
-            .find(|tensor| tensor.block_index == Some(3) && tensor.block_slot == Some(5))
-            .expect("source tensor")
-            .clone();
-        inv.tensors.push(duplicate);
-        refresh_inventory_derived_fields(&mut inv);
-
-        let err = validate_grok1_complete_manifest(&inv).unwrap_err();
-
-        assert!(format!("{err:#}").contains("duplicate structural name"));
-        assert!(format!("{err:#}").contains("block_003.slot_05"));
-    }
-
-    #[test]
-    fn grok1_coverage_fails_on_unexpected_key_layout() {
-        let mut inv = complete_grok1_inventory();
-        let tensor = inv
-            .tensors
-            .iter_mut()
-            .find(|tensor| tensor.block_index == Some(4) && tensor.block_slot == Some(3))
-            .expect("slot tensor");
-        tensor.kind = TensorKind::Unknown {
-            reason: "synthetic unexpected slot layout".to_string(),
-        };
-        refresh_inventory_derived_fields(&mut inv);
-
-        let err = validate_grok1_complete_manifest(&inv).unwrap_err();
-
-        assert!(format!("{err:#}").contains("unknown_tensors 1 != expected 0"));
-        assert!(format!("{err:#}").contains("unexpected tensor layout at block_004.slot_03"));
-    }
-
-    #[test]
-    fn grok1_coverage_checksum_is_reproducible() {
-        let inv = complete_grok1_inventory();
-
-        let first = validate_grok1_complete_manifest(&inv)
-            .expect("first manifest")
-            .checksum;
-        let second = validate_grok1_complete_manifest(&inv)
-            .expect("second manifest")
-            .checksum;
-
-        assert_eq!(first, second);
-    }
-
     fn shifted_grok_ckpt0_tensors() -> Vec<TensorInfo> {
         let mut tensors = Vec::new();
         tensors.push(tensor(
@@ -1320,150 +812,6 @@ mod tests {
         append_grok_blocks(&mut tensors, 2);
 
         tensors
-    }
-
-    fn complete_grok1_inventory() -> ModelInventory {
-        let mut tensors = vec![
-            complete_tensor(
-                0,
-                0,
-                None,
-                None,
-                TensorKind::TokenEmbedding,
-                TensorRole::Tensor,
-                TensorDType::F32,
-                vec![131_072, 6_144],
-            ),
-            complete_tensor(
-                1,
-                0,
-                None,
-                None,
-                TensorKind::FinalNorm,
-                TensorRole::Tensor,
-                TensorDType::F32,
-                vec![6_144],
-            ),
-        ];
-
-        for block in 0..64u32 {
-            for slot in 0..12u32 {
-                let shard = 2 + block * 12 + slot;
-                let (kind, role, dtype, shape) = match slot {
-                    0 | 2 => (
-                        TensorKind::MoeExpertProjection {
-                            projection: MoeProjection::Unresolved,
-                        },
-                        TensorRole::QuantWeight,
-                        TensorDType::I8,
-                        vec![8, 6_144, 32_768],
-                    ),
-                    1 => (
-                        TensorKind::MoeExpertProjection {
-                            projection: MoeProjection::Down,
-                        },
-                        TensorRole::QuantWeight,
-                        TensorDType::I8,
-                        vec![8, 32_768, 6_144],
-                    ),
-                    3 | 6 => (
-                        TensorKind::QuantizedAttentionProjection {
-                            width: QuantizedAttentionWidth::Narrow,
-                        },
-                        TensorRole::QuantWeight,
-                        TensorDType::I8,
-                        vec![6_144, 1_024],
-                    ),
-                    4 | 5 => (
-                        TensorKind::QuantizedAttentionProjection {
-                            width: QuantizedAttentionWidth::ModelWidth,
-                        },
-                        TensorRole::QuantWeight,
-                        TensorDType::I8,
-                        vec![6_144, 6_144],
-                    ),
-                    7 | 8 => (
-                        TensorKind::MoeScales,
-                        TensorRole::QuantScales,
-                        TensorDType::F32,
-                        vec![8, 32_768],
-                    ),
-                    9 | 10 => (
-                        TensorKind::BlockNorm,
-                        TensorRole::Tensor,
-                        TensorDType::F32,
-                        vec![6_144],
-                    ),
-                    11 => (
-                        TensorKind::Router,
-                        TensorRole::Tensor,
-                        TensorDType::F32,
-                        vec![6_144, 8],
-                    ),
-                    _ => unreachable!(),
-                };
-                tensors.push(complete_tensor(
-                    shard,
-                    0,
-                    Some(block),
-                    Some(slot),
-                    kind,
-                    role,
-                    dtype,
-                    shape,
-                ));
-            }
-        }
-
-        let mut inv = ModelInventory {
-            model_family: "grok-1".to_string(),
-            checkpoint_path: PathBuf::from("/tmp/grok-1/ckpt-0"),
-            shard_count: 770,
-            inferred: InferredHyperparams {
-                vocab_size: Some(131_072),
-                d_model: Some(6_144),
-                n_experts: Some(8),
-                d_ff: Some(32_768),
-                n_blocks: Some(64),
-            },
-            tensors,
-            blocks: Vec::new(),
-            totals: InventoryTotals::default(),
-            schema_version: SCHEMA_VERSION,
-        };
-        refresh_inventory_derived_fields(&mut inv);
-        inv
-    }
-
-    fn refresh_inventory_derived_fields(inv: &mut ModelInventory) {
-        inv.blocks = summarize_blocks(&inv.tensors);
-        inv.totals = compute_totals(&inv.tensors);
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn complete_tensor(
-        shard_ordinal: u32,
-        in_shard_index: u32,
-        block_index: Option<u32>,
-        block_slot: Option<u32>,
-        kind: TensorKind,
-        role: TensorRole,
-        dtype: TensorDType,
-        shape: Vec<u64>,
-    ) -> TensorInfo {
-        TensorInfo {
-            shard_path: PathBuf::from(format!("/tmp/grok-1/ckpt-0/tensor{shard_ordinal:05}_000")),
-            shard_ordinal,
-            in_shard_index,
-            role,
-            dtype,
-            shape: TensorShape::new(shape.clone()),
-            offset: 0,
-            nbytes: dtype.itemsize() as u64 * shape.iter().product::<u64>(),
-            kind,
-            block_index,
-            block_slot,
-        }
     }
 
     fn normal_grok_ckpt0_tensors_without_tail_norm() -> Vec<TensorInfo> {
